Move components to a component subfolder
